[datatype/value] Rename file value to datatype
